Multiplication logic implemented
Matrix multiplication logic has been implemented and is working.

HOWEVER: note that I added a data member and constructor in the "Consumer" class for testing purposes only. It adds a "Producer" data member so that I could test its matrices until the "SharedBuffer" is ready to go. Once that happens we can delete the "Producer" data member and all references to it (I've put comments where those references are) and we can switch to using the "SharedBuffer."