Added 'State' enum
'State' enum used to let us know if a workItem's subC matrix is 'READY' to be added to the original matrixC by the Producer, and to let us know if it is completely 'DONE' and no longer needed. Works pretty well as of now. Would like your input when we meet.

MatrixC is having issues printing. The subC matrices seem to be adding to it, and we can see the matrixC being printed in the console most of the time; however, when there are just a couple subC matrices left to calculate, matrixC stops printing. Need to fix.

- Allie